Create a consistent Connection API and separate RPC structs into individual files
