+++ .github/workflows/tests.yml
@@ -2,7 +2,7 @@ name: Test
 
 on:
   push:
-    branches: [main,connection-api]
+    branches: [main]
   pull_request:
     branches: [main]
 jobs:
