+++ CMakeLists.txt
@@ -13,10 +13,10 @@ conan_basic_setup()
 add_subdirectory(lib)
 
 # examples
-add_executable(example-get-account-info examples/GetAccountInfo.cpp)
-add_executable(example-account-subscribe examples/AccountSubscribe.cpp)
-add_executable(example-send-transaction examples/SendTransaction.cpp)
-add_executable(example-place-order examples/PlaceOrder.cpp)
+add_executable(example-get-account-info examples/get_account_info.cpp)
+add_executable(example-account-subscribe examples/account_subscribe.cpp)
+add_executable(example-send-transaction examples/send_transaction.cpp)
+add_executable(example-place-order examples/place_order.cpp)
 
 # link conan
 target_link_libraries(example-get-account-info ${CONAN_LIBS})
@@ -25,9 +25,9 @@ target_link_libraries(example-send-transaction ${CONAN_LIBS})
 target_link_libraries(example-place-order ${CONAN_LIBS})
 
 # libsol example
-add_executable(connection-example examples/ConnectionExample.cpp)
+add_executable(connection-example examples/connection_example.cpp)
 target_link_libraries(connection-example ${CONAN_LIBS})
-target_link_libraries(connection-example sol) # main libsol
+target_link_libraries(connection-example sol) # libsol
 
 # tests
 include(CTest)
