+++ include/solana.h
@@ -1,0 +1,20 @@
+#ifndef SOLCPP_SOLANA_H
+#define SOLCPP_SOLANA_H
+
+#include "solana/account.hpp"
+#include "solana/base58.hpp"
+#include "solana/base64.hpp"
+#include "solana/connection.hpp"
+#include "solana/fixedp.h"
+#include "solana/instruction.hpp"
+#include "solana/fixedp.h"
+#include "solana/instruction.hpp"
+#include "solana/int128.hpp"
+#include "solana/keypair.hpp"
+#include "solana/privatekey.hpp"
+#include "solana/pushcompact.hpp"
+#include "solana/request.hpp"
+#include "solana/transaction.hpp"
+#include "solana/types.hpp"
+
+#endif //SOLCPP_SOLANA_H
