+++ include/solana/request.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Types.h"
+#include "types.hpp"
 
 namespace sol {
     class Request{
