+++ lib/CMakeLists.txt
@@ -1,2 +1,2 @@
-include_directories(${solcpp_SOURCE_DIR}/include)
-add_library(sol Connection.cpp)
+include_directories(${solcpp_SOURCE_DIR}/include/solana)
+add_library(sol connection.cpp)
